fixengine: Add #resetOutgoingSeq in SessionStore
This method is added to support resetting of outgoing sequence number
when the FIX session is disconnected. If at some point a disconnected
session can be loaded from the store without the need of restablishing a
connection, this method can be dropped and replaced with #load and #save.

Also, implement it in MongoSessionStore.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@reaktor.fi>