fixengine: Validate presence of MsgSeqNum(34) during msg header parsing
After parsing of BeginString(8), BodyLength(9), MsgType(35), make sure
that MsgSeqNum(34) is present and valid. If not, despite of other errors
found in the message (with the exception of errors resulting the message
to be considered garbled), throw an exception indicating that
MsgSeqNum(34) is missing.

When a MsgSeqNum(34) is considered missing, MsgSeqNumMissingException is
thrown which in turn causes invocation of Callback#msgSeqNumMissing of
Parser. As recommended in FIX Protocol 4.3 vol. 2 (with errata September
20, 2002, page 30), the session is terminated when MsgSeqNum(34) is
missing.

While at it, update BUGS accordingly.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Kari Husa <kari.husa@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@reaktor.fi>