fixengine: Validate FIX messages only when the queue has no gaps
Currently, Session may end up rejecting messages out-of-order since
validation takes place immediately. For example, given that a queue has
gaps, that is received MsgSeqNum(34) > expected MsgSeqNum(34), then the
subsequent invalid message results into a rejection message, whereas the
rejection should be sent only after the gaps in the queue has been
filled.

The following exception applies to all messages:

 - received MsgSeqNum(34) does not exist in a message or is invalid
   resulting into a termination,

The following exceptions apply to messages other than Sequence Reset:

 - received MsgSeqNum(34) is smaller than the expected one resulting
   into a termination,

 - received MsgSeqNum(34) is greater than the expected one resulting
   into a resend request

The following exceptions apply to Sequence Reset - Gap Fill messages:

 - received MsgSeqNum(34) is smaller than the expected one resulting
   into a termination given that the PossDupFlag is not "Y" or not
   specified in the message

 - received MsgSeqNum(34) is smaller than the expected one causing the
   message to be ignored given that the PossDupFlag is "Y"

 - received MsgSeqNum(34) is greater then the expected one resulting
   into a resend request

The following exceptions apply to Sequence Reset - Reset messages:

 - received MsgSeqNum(34) is completely ignored and the message is
   processed without regards to the MsgSeqNum(34) of the message.

Although, MsgSeqNum(34) of Sequence Reset - Reset messages is not used,
it is required to be in the message and valid. If not specified or
valid, the session is terminated.

Also, some of the test cases have been fixed to use the correct sequence
number as the current version did not track the sequence numbers
correctly.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>