fixengine: Simplify detection of unexpected value of received MsgSeqNum(34)
After MsgSeqNum(34) has been parsed and the message is not a Sequence
Reset message, the very first thing before placing the message into the
queue is to verify that it is not out of sequence.

If an unexpected MsgSeqNum(34) does not result into a termination of a
session, it must be placed into the queue, and thus, all messages with
valid MsgSeqNum(34) values are enqueued, except the Sequence Reset -
Reset messages.

Note that the expected incoming MsgSeqNum(34) is not incremented until
the message is placed in the queue and only when the MsgSeqNum(34) is of
an expected value. Therefore, it is not required to save the expected
sequence number to a variable before enqueueing the received message.

If the message is received in order, the messages from the queue without
gaps is processed.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>