fixengine: Invoke Session#processOutOfSyncMessageQueue or #processInSyncMessageQueue
Currently, if the message queue is out of sync, #processOutOfSyncMessageQueue
is invoked after which #processInSyncMessageQueue is invoked, too. This
obviously is erroneous.

Note that if the received MsgSeqNum(34) is unexpected, the tracked
message sequence number in MessageQueue is not updated, and thus,
MessageQueue#nextSeqNum can be safely invoked instead of passing from
'expectedMsgSeqNum' from Session#processMessage.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Kari Husa <kari.husa@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>