fixengine: Test reconnecting with messages in outgoing queue
First of all, Logon requests are sent out-of-queue, since the initiator
must authenticated prior to sending queued messages when initiator
reconnects to the FIX acceptor.

As specified in FIX Protocol version 4.3, Volume 2, test case Ref ID 16
b, the acceptor must issue a ResendRequest since received MsgSeqNum(34)
of Logon message is greater than expected. At this point, initiator is
expected to resend queued messages with PossDupFlag(43) set to "Y".

Overload InitiatorSpecification#runInClient that receives a command that
is executed prior to opening a connection. This allows messages to be
queued but not sent.

While at it, check that connection is not null before accessing it in
Session#send.

Signed-off-by: Kari Husa <kari.husa@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>