fixengine: Establish session after receiving a valid response to a Logon Request
Currently, the session is not established after a response to a Logon
Request has been received if the MsgSeqNum(34) is not as expected. This
causes a Resend Request to be issued after which the other acceptor may
response with a lower MsgSeqNum(34) causing that response to be
processed before the response to the Logon Request.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>