fixengine: Allow conditionally required repeating groups
Also, declare repeating group starting with NoAllocs(78) of allocation
message as an optional one as defined in FIX 4.2 and introduce a test
case, where the repeating group is not present.

Note that RepeatingGroup#format and #toString return an empty string if
count tag is not present and therefore not parsed. This allows us to
consider the repeating group as not present in terms of whether the
count tag is present or not.

Update, AllocationMessage and BUGS accordingly.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>