fixengine: Filter out duplicate (PossResend) messages
If PossResend(96) = "Y", the message may be duplicate, as the
transmitting side cannot be sure that message has been delivered. In
order to verify the duplicity of the message, each incoming message is
stored and upon receiving a message with PossResend(96) = "Y", the
duplicity is checked using a user-defined rule as specified by
MessageComparator and retrieving previously stored messages.

Also, a test case as specified in the FIX protocol specification is
introduced and BUGS file is updated accordingly.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>