console-client: Unify command line arguments syntax
Now all arguments must be given in 'Name=value' syntax. This is done to
make it easier to remember the commands and to be able to implement tab
completion later.

Add helper class Arguments for parsing argument values.  Add examples
classes to test classpath to be able to write unit tests for them.

Signed-off-by: Markus Ylä-Ilomäki <markus.yla-ilomaki@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>