fixengine: Set Session.MAX_CONSECUTIVE_RESEND_REQUESTS to 10
Upon logging in, too low of a threshold may result into a logout that
would have been otherwise avoided.

For example, upon a successful login, if the other end replies with
another message in addition to a logon response and the MsgSeqNum(34) is
too high, two resend requests will be issued.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Signed-off-by: Pekka Enberg <penberg@kernel.org>