fix: Add DiskSessionStore
DiskSessionStore is a lightweight session store utilizing the disk
storage. Unlike the MongoDB session store, the disk session store
stores only the incoming and outgoing sequence numbers and not the
entire messages.

DiskSessionStore has one configuration parameter, the path. It
defines the directory in which the sessions are stored. Under it,
the session store creates a directory for each SenderCompID(49)
and TargetCompID(56) pair. Inside these directories, the session
store maintains a text file containing the incoming and outgoing
sequence numbers per each trading day. Given the path "/tmp", for
example, the directory structure looks like this:

  /tmp/fix
    /INITIATOR-ACCEPTOR1
      /2012-05-08
      /2012-05-09
    /INITIATOR-ACCEPTOR2
      /2012-05-08

The session tests can be run using the disk session store by
defining the system property "fixengine.session.store" to be
"disk". Additionally, "fixengine.session.store.path" can be used
to set the path, which defaults to "/tmp/fixengine". In SBT, a
system property can be set with the 'eval' command as follows:

  eval System.setProperty("fixengine.session.store", "disk")

Some of the session tests expect the session store to be capable
to store the incoming and outgoing messages. As the disk session
store does not do that, the following session tests fail with it:

  - RecvAppOrAdminMsgSpec
  - TestPossResendHandlingSpec
  - RecvResendRequestSpec

This will be fixed at a later date.

Signed-off-by: Jussi Virtanen <jussi.virtanen@valotrading.com>