fix: Make FloatField locale independent
FloatField uses the default locale for the decimal separator when
converting a number value into text format. If the default locale
is set to, for example, Finnish, the decimal separator is a comma
(',') rather than a full stop ('.').

As the protocol demands a full stop, fix this by explicitly using
the US decimal format symbols.

Signed-off-by: Jussi Virtanen <jussi.virtanen@valotrading.com>