itch: Add ByteString#getBytesUnsafe
This method can be used to access the internal bytes of a ByteString. It's
called `getBytesUnsafe` because it allows access into the internal state, and
caller must be careful what to do with it. In some cases this is, however,
required.

In case the ByteBuffer is a complete buffer, `getBytesUnsafe` just directly
returns the byte array. In all other cases, it returns a copy of the part the
ByteString is pointing to.

Signed-off-by: Aki Saarinen <aki.saarinen@valotrading.com>