Add NASDAQ OMX SoupTCP 1.0 support
The NASDAQ OMX SoupTCP 1.0 implementation comprises of a SoupTCP client
and a SoupTCP server. They handle link failure detection, but sequence
numbers must handled on a higher level.

The test environment introduces a new test framework based on ScalaTest
that aims to make it easier to test Silvertip event sources.

Signed-off-by: Jussi Virtanen <jussi.virtanen@valotrading.com>